:recycle: Use the same kind of test setup as the runtime
