+++ _integration_tests/src/fixture_providers/index.ts
@@ -1,0 +1,2 @@
+export * from './process_instance_handler';
+export * from './test_fixture_provider';
