+++ _integration_tests/src/index.ts
@@ -1,3 +1,2 @@
-export * from './iam_service_mock';
-export * from './process_instance_handler';
-export * from './test_fixture_provider';
+export * from './fixture_providers/index';
+export * from './mocks/index';
