+++ _integration_tests/src/mocks/index.ts
@@ -1,0 +1,1 @@
+export * from './iam_service_mock';
