+++ _integration_tests/test/process_models/get_process_model_events.js
@@ -2,7 +2,7 @@
 
 const should = require('should');
 
-const TestFixtureProvider = require('../../dist/commonjs/test_fixture_provider').TestFixtureProvider;
+const TestFixtureProvider = require('../../dist/commonjs').TestFixtureProvider;
 
 describe('Management API:   GET  ->  /process_models/:process_model_id/events', () => {
 
@@ -25,7 +25,7 @@ describe('Management API:   GET  ->  /process_models/:process_model_id/events', 
 
     const processModel = await testFixtureProvider
       .managementApiClientService
-      .getEventsForProcessModel(testFixtureProvider.identity, processModelId);
+      .getEventsForProcessModel(testFixtureProvider.identities.defaultUser, processModelId);
 
     should(processModel).have.property('events');
     should(processModel.events).be.instanceof(Array);
